fix / anime 要素の読み込み機能の仕様変更
数値を読み取る際、整数ではなく、実数型で読み込むよう変更

+++ Assets/Scenes/scripts/loaders/AnimeElementConverter.cs
@@ -31,7 +31,7 @@
 
                         if (propInfo != null)
                         {
-                            propInfo.SetValue(anime, int.Parse(attribute.Value));
+                            propInfo.SetValue(anime, double.Parse(attribute.Value));
                         }
                         else
                         {
